backend/requirements.txt
# Proxy AI Backend
fastapi==0.104.1
uvicorn==0.24.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
elevenlabs>=1.0.0